fix: added ssh private key from env

+++ packages/core/src/instance/instance.ts
@@ -4,19 +4,18 @@ import { existsSync } from "fs";
 import { dirname, join } from "path";
 import { env } from "../utils/env.js";
 import { InstanceCmdFactories } from "./instanceFactories.js";
-import { polling } from "./utils.js";
+import { createPrivateKeyFile, polling } from "./utils.js";
 import { fileURLToPath } from "url";
 
 type InstanceConfigType = {
   region?: string;
-  sshPrivateKey?: string;
-  identityFilePath?: string;
   tempDir?: string;
+  sshPrivateKey: string;
+  identityFilePath: string;
 };
 
 type LaunchInstanceConfig = {
   name: string;
-  sshPublicKey: string;
   instanceType?: string;
   imageId?: string;
   imageType?: string;
@@ -39,12 +38,17 @@ export class Instance implements IInstance {
   private launchedInstanceId: string | null;
   private instanceName: string | null;
   private semaphore: string;
-  private identityFilePath: string;
+  private privateFile: string;
   private publicDns: string | null;
   private liveUrl: string | null;
   private cmd: typeof InstanceCmdFactories;
+  private sshPrivateKey: string;
 
-  constructor({ region = "us-east-1", identityFilePath }: InstanceConfigType) {
+  constructor({
+    region = "us-east-1",
+    identityFilePath,
+    sshPrivateKey,
+  }: InstanceConfigType) {
     this.cmd = InstanceCmdFactories;
 
     this.semaphore = "/etc/prbranch/ready";
@@ -52,8 +56,12 @@ export class Instance implements IInstance {
     this.instanceName = null;
     this.publicDns = null;
     this.liveUrl = null;
-    this.identityFilePath = identityFilePath || join(process.cwd(), "private");
-
+    this.sshPrivateKey = sshPrivateKey;
+    this.privateFile = createPrivateKeyFile(
+      sshPrivateKey,
+      `${identityFilePath}`,
+      "private.key"
+    );
     this.client = InstanceCmdFactories.createInstance({
       region,
     });
@@ -166,9 +174,22 @@ export class Instance implements IInstance {
   // stop docker
   private async stopDockerContainer() {
     //node thinks one backslash as escape
-    await this.ssh(`docker stop \\$(docker ps -aq)`);
-    await this.ssh(`docker rm \\$(docker ps -aq)`);
-    await this.ssh(`docker rmi \\$(docker images -q)`);
+    const runningContainerId = await this.getRunningContainer();
+    if (runningContainerId) {
+      await this.ssh(`docker stop ${runningContainerId}`);
+      await this.ssh(`docker rm ${runningContainerId}`);
+      await this.ssh(`docker rmi \\$(docker images -q)`);
+    }
+  }
+
+  private async getRunningContainer() {
+    try {
+      const container = await this.ssh(`docker ps -aq`);
+      console.log({ containerId: container });
+      container.stdout;
+    } catch (error) {
+      return null;
+    }
   }
 
   async waitUntilInstance() {
@@ -209,7 +230,6 @@ export class Instance implements IInstance {
       await this.scp({
         source: sourcePath,
         target: serverAppPath,
-        mode: "0600",
       });
     } catch (e) {
       console.error(e);
@@ -243,20 +263,10 @@ export class Instance implements IInstance {
     return this.liveUrl;
   }
 
-  private async scp({
-    source,
-    target,
-    fileName,
-    mode = "0600",
-  }: {
-    source: string;
-    target: string;
-    fileName?: string;
-    mode?: string;
-  }) {
+  private async scp({ source, target }: { source: string; target: string }) {
     const host = this.publicDns;
     const remoteUser = "ec2-user";
-    const tempPrivateKey = this.identityFilePath;
+    const tempPrivateKey = this.privateFile;
 
     console.log("Starting to cpy files to server 📁 ---> 📂");
 
@@ -294,40 +304,38 @@ export class Instance implements IInstance {
       });
   }
 
-  private async ssh(cmd: string, file?: string, debug = false) {
+  private async ssh(cmd: string, debug = false) {
     const publicDns = this.publicDns;
-    const sshAddress = `ec2-user@${publicDns}`;
-    const tempPrivateKey = this.identityFilePath;
+    const sshHostAddress = `ec2-user@${publicDns}`;
+    const tempPrivateKey = this.privateFile;
 
-    const cmdToRun = `ssh ${
-      debug ? "-vvv" : ""
-    } -o StrictHostKeyChecking=no -o ServerAliveInterval=15 -o UserKnownHostsFile=/dev/null -o LogLevel=ERROR -o ConnectTimeout=10 -i "${tempPrivateKey}" "${sshAddress}" "${cmd}"`;
+    const sshDefaultOptions = [
+      "-o StrictHostKeyChecking=no",
+      "-o ServerAliveInterval=15",
+      "-o UserKnownHostsFile=/dev/null",
+      "-o LogLevel=ERROR",
+      "-o ConnectTimeout=10",
+    ];
 
-    const cpyFile = `cat ${file} | ${cmdToRun}`;
-    console.log(`\n running cmd..`);
-    console.log(`${file ? cpyFile : cmdToRun}\n`);
+    debug ? sshDefaultOptions.push("-vvv") : "";
 
-    if (file) {
-      return await $({ verbose: true })`${cpyFile}`
-        .then((_) => {
-          console.log(`\n successfully ran this cmd ${cpyFile} \n`);
-        })
-        .catch((e) => {
-          console.error(e);
-          throw new Error("Error while logging the cmd with file ", {
-            cause: e,
-          });
-        });
-    } else {
-      return await $({ verbose: true, shell: true })`${cmdToRun}`
-        .then((_) => {
-          console.log(`\n successfully ran this cmd ${cmdToRun} \n`);
-        })
-        .catch((e) => {
-          console.error(e);
-          throw new Error("Error while logging the cmd", { cause: e });
-        });
-    }
+    const cmdToRun = `ssh ${sshDefaultOptions.join(
+      " "
+    )} -i "${tempPrivateKey}" "${sshHostAddress}" "${cmd}"`;
+
+    console.log("\n----\n");
+    console.log(`\n running cmd..\n`, cmdToRun);
+    console.log("\n----\n");
+
+    return await $({ verbose: true, shell: true })`${cmdToRun}`
+      .then((res) => {
+        console.log(`\n successfully ran this cmd ${cmdToRun} \n`);
+        return res;
+      })
+      .catch((e) => {
+        console.error(e);
+        throw new Error("Error while logging the cmd", { cause: e });
+      });
   }
 
   private updateInstanceState(currentInstance: InitializeInstance) {
